requirements.txt
discord.py>=2.3,<3
python-dotenv>=1,<2
gspread>=6,<7
google-auth>=2,<3
